We now allow XOR options to have requirements!
What this means is that even mutually exclusive options can have
requirements, which will only be enforced /if/ that version of the XOR options
is actually present on the command line.

Example: it allows one to say that IF this option is present it should
appear exactly twice.

Given this configuration of a mutually exclusive or group of options:

    cmd.addXOR(
        option(long_name("file"), collect<string>(), at_most(1)),
        option(collect<string>(), exactly(2))
    );

Only the following command line(s) are acceptable:

    ...
        meaning: no option "--file ARG" nor an ARG(ument)
    .... --file ARG ...
        the --file option may appear exactly once only
    ...  ARG ARG ...
        IF an argument is specified, there HAVE to be two

Any mixture or repeating is greeted with a mighty parse error and fatal
termination of the program.