Can now have "match(std::regex const&)" w/o compile error
The library assumed the argument(s) passed to "match(...)", fed directly to
the constructor of std::regex, were always supporting "std::ostream<<(...)"
(specifically the first one).
This meant that if a pre-compiled std::regex was fed, it would bunk out.

Now you can. But the "human readable form" of the compiled regex will read
"compiled std::regex" so the author should document the pattern him/herself
in a docstring for the option.