Add the required() requirement predicate, mostly for XOR groups
In order to support marking a particular mutually-exclusive-or group as
required in a sensible way, this predicate had to be added.

It would be easy to allow any of the predicates ("exactly(n)", "at_least(n)"
&cet) but that wouldn't make any logical sense:
marking an exclusive-or group of options should mean one thing only: exactly
one of the alternatives SHOULD be given on the command line, no more, no
less.

The "required()" predicate can be added to a normal command line option's
definition where it will act exactly as "exactly(1)" - i.e. requiring the
command line option to be present exactly once, exactly as expected.