fixed problem with realpath()
The parse() method used "realpath()" to find out the name of the executable
itself (the program name). However that call failed if the directory
containing the program was added to the PATH variable and the program was
just called w/o leading absolute or relative path. Then realpath() returns
null which triggered a fatal error.