Added support for hidden command line options
You can now add the argparse::hidden() option to any option to make it
hidden; it will not show up under the automatically generated help.

If the option is parsed from the command line all constraints &cet are still
verified. If the userinput fails to meet (at least one of) the constraints
we DO print out the whole documentation of the option - obviously the
(expert?) user is aware this option exists, so there's no point in hiding it
any more; cat out of bag and all, and thus he or she may usefully be
informed where he/she went wrong.

An attempt is made to prevent constructing required invisible options - that
would be mighty silly and confusing to the user: something is required but
the framework is not allowed to tell what; it just won't stand for that.

Unfortunately most of the checks can only be done at runtime in stead of
compile time (at least not without giving up /some/ freedom).